requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
duckdb>=1.3
pyproj>=3.7
pyyaml>=6.0
